refactor : consolidate email and user API functions into single files and update imports
